NDEV-1149 Use naersk instead of cargo2nix
With cargo2nix we can use `nix build`, but we'd have to maintain a
Cargo.nix file which would have to be updated whenever Cargo.lock is
updated, and maintaining that would be a problem.

`nix build` is not working with naersk, but the main goal of this
flake is to have a `nix develop` to quickly have a Rust toolchain and
call cargo (and other tools such as rust-analyzer).